Atomic components folder structure + navigation
- Refactor folder structure to follow atomic components style.
  This is done in order to have DRY code and clean organization practices.
  See: https://codeburst.io/atomic-design-with-react-e7aea8152957
  for additional context.
  It's worth to note that this practice can be followed as strictly as we
  like; for example, if an organism requires an atom that no other
  resource will use, it may be best to simply add the code in the organism
  itself.

- Add navigation layout for all main pages
