+++ src/components/templates/layout.js
@@ -1,0 +1,23 @@
+import React from "react"
+import PropTypes from "prop-types"
+import { useStaticQuery, graphql } from "gatsby"
+
+import Header from "../organisms/header"
+
+const Layout = () => {
+  const data = useStaticQuery(graphql`
+    query SiteTitleQuery {
+      site {
+        siteMetadata {
+          title
+        }
+      }
+    }
+  `)
+
+  return (
+    <Header siteTitle={data.site.siteMetadata.title} />
+  )
+}
+
+export default Layout
